+++ src/ldc/filter/__init__.py
@@ -27,6 +27,7 @@ from ._skip_duplicate_ids import SkipDuplicateIDs
 from ._skip_duplicate_text import SkipDuplicateText
 from ._split_records import SplitRecords
 from ._strip_strings import StripStrings
+from ._sub_process import SubProcess
 from ._tee import Tee
 from ._text_length import TextLength
 from ._text_stats import TextStatistics